+++ crates/volta-core/src/lib.rs
@@ -20,5 +20,3 @@ pub mod sync;
 pub mod tool;
 pub mod toolchain;
 pub mod version;
-
-const VOLTA_FEATURE_PNPM: &str = "VOLTA_FEATURE_PNPM";