+++ crates/volta-core/src/tool/mod.rs
@@ -1,4 +1,3 @@
-use std::env;
 use std::fmt::{self, Display};
 use std::path::PathBuf;
 
@@ -8,7 +7,6 @@ use crate::session::Session;
 use crate::style::{note_prefix, success_prefix, tool_version};
 use crate::sync::VoltaLock;
 use crate::version::VersionSpec;
-use crate::VOLTA_FEATURE_PNPM;
 use cfg_if::cfg_if;
 use log::{debug, info};
 
@@ -91,27 +89,14 @@ impl Spec {
                 None => Ok(Box::new(BundledNpm)),
             },
             Spec::Pnpm(version) => {
-                // If the pnpm feature flag is set, use the special-cased package manager logic
-                // to handle resolving (and ultimately fetching / installing) pnpm. If not, then
-                // fall back to the global package behavior, which was the case prior to pnpm
-                // support being added
-                if env::var_os(VOLTA_FEATURE_PNPM).is_some() {
-                    let version = pnpm::resolve(version, session)?;
-                    Ok(Box::new(Pnpm::new(version)))
-                } else {
-                    let package = Package::new("pnpm".to_owned(), version)?;
-                    Ok(Box::new(package))
-                }
+                let version = pnpm::resolve(version, session)?;
+                Ok(Box::new(Pnpm::new(version)))
             }
             Spec::Yarn(version) => {
                 let version = yarn::resolve(version, session)?;
                 Ok(Box::new(Yarn::new(version)))
             }
-            // When using global package install, we allow the package manager to perform the version resolution
-            Spec::Package(name, version) => {
-                let package = Package::new(name, version)?;
-                Ok(Box::new(package))
-            }
+            Spec::Package(name, version) => Ok(Box::new(Package::new(name, version))),
         }
     }
 
@@ -129,16 +114,10 @@ impl Spec {
                 feature: "Uninstalling npm".into(),
             }
             .into()),
-            Spec::Pnpm(_) => {
-                if env::var_os(VOLTA_FEATURE_PNPM).is_some() {
-                    Err(ErrorKind::Unimplemented {
-                        feature: "Uninstalling pnpm".into(),
-                    }
-                    .into())
-                } else {
-                    package::uninstall("pnpm")
-                }
+            Spec::Pnpm(_) => Err(ErrorKind::Unimplemented {
+                feature: "Uninstalling pnpm".into(),
             }
+            .into()),
             Spec::Yarn(_) => Err(ErrorKind::Unimplemented {
                 feature: "Uninstalling yarn".into(),
             }