+++ src/command/list/human.rs
@@ -178,8 +178,8 @@ fn display_packages(packages: &[Package]) -> String {
         String::from(
             "⚡️ No tools or packages installed.
 
-You can safely install packages by running `volta install <package name>`.
-See `volta help install` for details and more options.",
+Install global packages with your package manager, for example
+`pnpm add --global <package name>`.",
         )
     } else {
         format!(
@@ -195,8 +195,8 @@ fn display_tool(tool: &str, host_packages: &[Package]) -> String {
         format!(
             "⚡️ No tools or packages named `{}` installed.
 
-You can safely install packages by running `volta install <package name>`.
-See `volta help install` for details and more options.",
+Install global packages with your package manager, for example
+`pnpm add --global <package name>`.",
             tool
         )
     } else {
@@ -1136,8 +1136,8 @@ See `volta help install` for details and more options.";
         fn none() {
             let expected = "⚡️ No tools or packages installed.
 
-You can safely install packages by running `volta install <package name>`.
-See `volta help install` for details and more options.";
+Install global packages with your package manager, for example
+`pnpm add --global <package name>`.";
 
             assert_eq!(display_packages(&[]), expected);
         }
@@ -1262,8 +1262,8 @@ See `volta help install` for details and more options.";
         fn none() {
             let expected = "⚡️ No tools or packages named `ember` installed.
 
-You can safely install packages by running `volta install <package name>`.
-See `volta help install` for details and more options.";
+Install global packages with your package manager, for example
+`pnpm add --global <package name>`.";
 
             assert_eq!(display_tool("ember", &[]), expected);
         }