+++ tests/acceptance/volta_pin.rs
@@ -928,7 +928,6 @@ fn pin_pnpm_no_node() {
         .package_json(BASIC_PACKAGE_JSON)
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -949,7 +948,6 @@ fn pin_pnpm() {
         .package_json(&package_json_with_pinned_node("1.2.3"))
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -970,7 +968,6 @@ fn pin_pnpm_reports_info() {
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
         .env(VOLTA_LOGLEVEL, "info")
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -987,7 +984,6 @@ fn pin_pnpm_latest() {
         .package_json(&package_json_with_pinned_node("1.2.3"))
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -1007,7 +1003,6 @@ fn pin_pnpm_no_version() {
         .package_json(&package_json_with_pinned_node("1.2.3"))
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -1025,7 +1020,6 @@ fn pin_pnpm_no_version() {
 fn pin_pnpm_missing_release() {
     let s = sandbox()
         .package_json(&package_json_with_pinned_node("1.2.3"))
-        .env("VOLTA_FEATURE_PNPM", "1")
         .mock_not_found()
         .build();
 
@@ -1050,7 +1044,6 @@ fn pin_node_and_pnpm() {
         .distro_mocks::<NodeFixture>(&NODE_VERSION_FIXTURES)
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -1070,7 +1063,6 @@ fn pin_pnpm_leaves_npm() {
         .package_json(&package_json_with_pinned_node_npm("1.2.3", "3.4.5"))
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(