+++ tests/acceptance/volta_run.rs
@@ -440,7 +440,6 @@ fn command_line_pnpm() {
         .pnpm_available_versions(PNPM_VERSION_INFO)
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
         .env(VOLTA_LOGLEVEL, "debug")
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -460,7 +459,6 @@ fn inherited_pnpm() {
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
         .package_json(&package_json_with_pinned_node_pnpm("10.99.1040", "7.7.1"))
         .env(VOLTA_LOGLEVEL, "debug")
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(
@@ -480,7 +478,6 @@ fn force_no_pnpm() {
         .distro_mocks::<PnpmFixture>(&PNPM_VERSION_FIXTURES)
         .package_json(&package_json_with_pinned_node_pnpm("10.99.1040", "7.7.1"))
         .env(VOLTA_LOGLEVEL, "debug")
-        .env("VOLTA_FEATURE_PNPM", "1")
         .build();
 
     assert_that!(